+++ cdot65/prisma_access/plugins/modules/address_groups.py
@@ -1,5 +1,5 @@
 """
-Ansible module for managing tags in Prisma Access.
+Ansible module for managing address objects in Prisma Access.
 Copyright: (c) 2023, Calvin Remsburg (@cdot65) <cremsburg.dev@gmail.com>
 """
 from __future__ import absolute_import, division, print_function
@@ -16,7 +16,7 @@ from ansible_collections.cdot65.prisma_access.plugins.module_utils.api_spec impo
 
 # Prisma Access SDK
 from panapi import PanApiSession
-from panapi.config.objects import Tag
+from panapi.config.objects import AddressGroup
 
 # jwt is not a float and causes an error of the token not being valid yet, ugh
 import time
@@ -25,40 +25,45 @@ __metaclass__ = type
 
 DOCUMENTATION = r"""
 ---
-module: tag
+module: address groups
 
-short_description: Manage tag objects.
+short_description: Manage address group objects.
 
-version_added: "0.1.0"
+version_added: "0.1.2"
 
-description: Manage tag objects within Prisma Access.
+description: Manage address group objects within Prisma Access.
 
 options:
-    token:
+    description:
         description:
-            - used to authenticate to the API
-        required: true
+            - Description of the address object.
+        required: false
         type: str
-    folder:
+    dynamic:
         description:
-            - The folder you would like to associate to these tags goes here
+            - declare the address group object is dynamic
+        required: false
+        type: dict
+        options:
+            filter:
+                required: True
+                type: str
+    folder:
         choices:
-          - 'GlobalProtect'
-          - 'Mobile Users'
-          - 'Remote Networks'
-          - 'Service Connections'
-          - 'Shared'
+          - "Shared"
+          - "Mobile Users"
+          - "Remote Networks"
+          - "Service Connections"
+          - "Mobile Users Container"
+          - "Mobile Users Explicit Proxy"
+        description:
+            - declare where the object should reside.
         required: true
         type: str
     name:
         description:
-            - Value of the tag's name
+            - Value of the address group object's name
         required: true
-        type: str
-    comments:
-        description:
-            - Additional comments about the tag
-        required: False
         type: str
     state:
         description:
@@ -68,54 +73,75 @@ options:
           - 'absent'
           - 'present'
         type: str
+    static:
+        description:
+            - declare whether the address group object is static
+        required: false
+        type: list
+
 
 extends_documentation_fragment:
-    - cdot65.prisma_access.tag
+    - cdot65.prisma_access.address_groups
 
 author:
     - Calvin Remsburg (@cdot65)
 """
 
 EXAMPLES = r"""
-    - name: Create tags
-      cdot65.prisma_access.tag:
+    - name: Create address group
+      cdot65.prisma_access.address_groups:
         provider:
-          client_id: "automation@1111111111.iam.panserviceaccount.com"
-          client_secret: "11111111-1111-1111-1111-111111111111"
-          scope: "1111111111"
-        name: "{{ item.name }}"
-        color: "{{ item.color }}"
-        comments: "{{ item.comments }}"
+          client_id: "{{ client_id }}"
+          client_secret: "{{ client_secret }}"
+          scope: "{{ scope }}"
+        name: "AnsibleTestGroupStatic"
         folder: "Service Connections"
+        description: "This is just a test"
+        static:
+          - "AnsibleTestAddress"
+        tag:
+          - "ansible-test"
         state: "present"
-      loop:
-        - name: "ansible-test"
-          color: "Lavender"
-          comments: "This is a test from Ansible"
+
+    - name: Create address group
+      cdot65.prisma_access.address_groups:
+        provider:
+          client_id: "{{ client_id }}"
+          client_secret: "{{ client_secret }}"
+          scope: "{{ scope }}"
+        name: "AnsibleTestGroupDynamic"
+        folder: "Service Connections"
+        description: "This is just a test"
+        dynamic:
+          filter: "'ansible-test'"
+        tag:
+          - "ansible-test"
+        state: "present"
+
 """
 
 
 def main():
     """
     The main function is the entry point for the script.
-    It performs actions to create or delete a tag based on the state parameter.
+    It performs actions to create or delete an address based on the state parameter.
 
     Summary of the main function:
       - sets up the required parameters
       - creates the necessary objects
       - performs actions on the objects.
 
-    If the `state` parameter is set to "absent": the function will delete the tag (if it exists).
+    If the `state` parameter is set to "absent": the function will delete the address (if it exists).
 
-    If the `state` parameter is set to anything else: the function will create the tag (if it does not exist).
+    If the `state` parameter is set to anything else: the function will create the address (if it does not exist).
 
-    If the tag already exists or is successfully deleted, the function will exit with a changed value of False.
+    If the address already exists or is successfully deleted, the function will exit with a changed value of False.
 
-    If the tag is successfully created or deleted, the function will exit with a changed value of True.
+    If the address is successfully created or deleted, the function will exit with a changed value of True.
 
     The function may raise an Exception and will fail if an error occurs.
     """
-    module = AnsibleModule(argument_spec=PrismaAccessSpec.tag_spec())
+    module = AnsibleModule(argument_spec=PrismaAccessSpec.address_groups_spec())
 
     try:
         auth = module.params.get("provider")
@@ -130,25 +156,38 @@ def main():
         # jwt isn't a float, causing an error of the token not being valid yet
         time.sleep(1)
 
-        tag = Tag(
+        group = AddressGroup(
+            description=module.params["description"],
             folder=module.params["folder"],
             name=module.params["name"],
-            comments=module.params["comments"],
-            color=module.params["color"],
             state=module.params["state"],
         )
 
-        already_exists = False
-        existing_tags = tag.list(session)
+        if module.params["tag"]:
+            group.__setattr__("tag", module.params["tag"])
 
-        for each in existing_tags:
-            if tag.name == each.name:
+        if module.params["static"]:
+            group.__setattr__("static", module.params["static"])
+        elif module.params["dynamic"]:
+            group.__setattr__("dynamic", module.params["dynamic"])
+        else:
+            module.fail_json(
+                msg="Must define either static or dynamic address group"
+            )
+
+        already_exists = False
+        existing_address_groups = group.list(session)
+
+        for each in existing_address_groups:
+            if group.name == each.name:
                 already_exists = True
-                tag.id = each.id
+                group.id = each.id
+
+        # raise Exception(already_exists)
 
         if module.params["state"] == "absent":
             if already_exists is True:
-                tag.delete(session)
+                group.delete(session)
                 if session.response.status_code != 200:
                     module.fail_json(
                         msg=f"Did not receive proper response: {session.response.text}"
@@ -159,12 +198,12 @@ def main():
                 )
             else:
                 module.exit_json(
-                    changed=False, data="Tag does not exist, exiting"
+                    changed=False, data="Group does not exist, exiting"
                 )
 
         else:
             if already_exists is False:
-                tag.create(session)
+                group.create(session)
                 if session.response.status_code != 201:
                     module.fail_json(
                         msg=f"Did not receive proper response: {session.response.text}"