+++ cdot65/prisma_access/CHANGELOG.rst
@@ -5,6 +5,34 @@ cdot65.prisma_access Release Notes
 .. contents:: Topics
 
 
+v0.1.2
+======
+
+Release Summary
+---------------
+
+Adds support for address group objects.
+
+
+New Modules
+-----------
+
+- cdot65.prisma_access.address_groups - Creates or removes address group objects within Prisma Access.
+
+v0.1.1
+======
+
+Release Summary
+---------------
+
+This corrects authentication and adds support for address objects.
+
+
+New Modules
+-----------
+
+- cdot65.prisma_access.addresses - Creates or removes address objects within Prisma Access.
+
 v0.1.0
 ======